Set Anchor Cells from Spatial Vector Objects

Returns a vector of cell numbers overlapping with the locations of spatial vector data. These cells can be used as anchor cells in other scapesClassification functions.

anchor.svo(
  rstack,
  spatial_vector_name,
  only_NAs = FALSE,
  fill_NAs = FALSE,
  plot = TRUE,
  saveRDS = NULL,
  writeRaster = NULL,
  overWrite = FALSE
)

Arguments

rstack

Raster* object.

spatial_vector_name

OGR data source name.

only_NAs

logic, return only cell numbers overlapping with the spatial vector data that have missing values in rstack.

fill_NAs

logic, return cell numbers of cells adjacent to those overlapping with the spatial vector data that have missing values in rstack.

plot

logic, plot anchor points.

saveRDS

filename, if a file name is provided save the anchor cell vector as an RDS file.

writeRaster

filename, if a raster name is provided save the anchor cell vector as a raster file.

overWrite

logic, if RDS and raster names already exist, existing files are overwritten.

Value

anchor cell vector.

Details

When the arguments only_NA and fillNAs are FALSE the output is equivalent to the output of the function raster::extract() over a raster whose values are cell numbers. Cell numbers start with 1 in the upper-left corner and increase from left to right and from top to bottom.