Test Conditions for All Cells

Evaluate conditions for cells that have not already been classified and classify them if conditions are true.

cond.4.all(
  attTbl,
  conditions,
  classVector = NULL,
  class,
  overwrite_class = FALSE
)

Arguments

attTbl

data.frame, the attribute table returned by the function attTbl.

conditions

character string, the conditions a cell have to meet to be classified as indicated by the argument class. Absolute conditions can be used (see conditions).

classVector

numeric vector, defines the cells in the attribute table that have already been classified.

class

numeric, the numeric class to attribute to the cells meeting conditions.

overwrite_class

logic, reclassify cells that have already been classified.

Value

Update classVector with the new cells that were classified by the function. If no classVector was provided, the function return a new classification vector.

Details

The function evaluates the conditions of the argument conditions for all cells that have not already been classified. The cells for which conditions are TRUE are classified as indicted by the argument class.

See also