Test Conditions for Neighbors and Neighbors of Neighbors

Evaluate conditions for cells neighboring specific classes and classify them if conditions are TRUE.

cond.4.nofn(
  attTbl,
  ngbList,
  nbsIndex = FALSE,
  classVector,
  nbs_of,
  conditions,
  class,
  min.border = NULL,
  overwrite_class = FALSE,
  max.iter = +Inf,
  fn_perc = 1,
  directional = T
)

Arguments

attTbl

data.frame, the attribute table returned by the function attTbl.

ngbList

list, it has to contain the list of 8-neighbors of each cell in attTbl$Cell (see ngbList).

nbsIndex

logic, ngbList contain the neighbors index position in the attribute table (see ngbList).

classVector

numeric vector, defines the cells in the attribute table that have already been classified.

nbs_of

numeric or numeric vector, indicates the neighbors of the classes that have to be evaluated. If it includes the class of the argument class, the function will evaluate the conditions also for the neighbors of the neighbors.

conditions

character string, the conditions a cell have to meet to be classified as indicated by the argument class. Absolute, focal cell and focal neighborhood conditions can be used. Condition strings can include only one neighborhood condition (see conditions).

class

numeric, the numeric class to attribute to the cells meeting conditions.

min.border

numeric value between 0 and 1. It indicates the minimum percentage of cells adjacent to the cell in evaluation that have to belong to one of the classes in nbs_of. If one, then all neighboring cells have to belong to one of the classes in nbs_of for the cell in evaluation being classified as class.

overwrite_class

logic, reclassify cells that have already been classified.

max.iter

integer, the maximum number of iterations.

fn_perc

numeric value between 0 and 1. If neighborhood conditions are considered, it determines the percentage of cells in the neighborhood for which the conditions have to be true in order to classify the cell being evaluated as indicted by the argument class (see conditions).

directional

logic, only directional neighbors are considered to test neighborhood conditions. The argument fn_perc will also consider only directional neighbors (see conditions).

Value

Update classVector with the new cells that were classified by the function.

Details

The function evaluates the conditions of the argument conditions for all cells neighboring cells classified as one of the classes included in nbs_of. If the argument nbs_of includes the class of the argument class, then at each iteration the function will evaluate if among the neighbors of the newly classified cells there are cells meeting conditions and it will classify them accordingly.

See also