Test Conditions and Reclassify

Evaluate conditions for cells of a class and reclassify them if conditions are true.

cond.reclass(
  attTbl,
  ngbList = NULL,
  nbsIndex = FALSE,
  classVector,
  conditions,
  class,
  reclass,
  fn_perc = 1
)

Arguments

attTbl
data.frame, the attribute table returned by the function attTbl.

ngbList
list, it has to contain the list of 8-neighbors of each cell in attTbl$Cell (see ngbList). If conditions do not include focal neighborohood conditions this argument can be NULL (see conditions).

nbsIndex
logic, ngbList contain the neighbors index position in the attribute table (see ngbList).

classVector
numeric vector, defines the cells in the attribute table that have already been classified.

conditions
character string, the conditions a cell have to meet to be classified as indicated by the argument reclass. Absolute and focal neighborhood conditions can be used (see conditions).

class
numeric or numeric vector, indicates the classes for which conditions have to be evaluated.

reclass
numeric, the numeric class to attribute to the cells meeting conditions.

fn_perc
numeric value between 0 and 1. If a focal neighborhood condition is considered, it determines the percentage of cells in the focal neighborhood for which the conditions have to be TRUE in order to classify the cell being evaluated as indicted by the argument reclass (see conditions).

Value

Update classVector with the new cells that were classified by the function.

Details

The function evaluates the conditions of the argument conditions for all cells in the classes of the argument class. The cells for which conditions are true are reclassified as indicted by the argument reclass.

See also