Check Conditions

Check for spelling and syntax errors in conditions and evaluate the type of conditions being used.

conditions(names_attTbl, cond, silent = FALSE)

Arguments

names_attTbl

character vector, the column (i.e. variable) names of the attribute table returned by the function attTbl.

cond

character string, a set of conditions used to classify raster cells (see cond.4.nofn, cond.reclass, cond.4.all and anchor.seed).

silent

logic, returns only error messages.

Value

An error message as character string if spelling or syntax errors were found or a message detailing the types of conditions that were found.

Details

Conditions

Conditions are passed to scapesClassification functions as a single character string. They can consist of combination of arithmetic (+|-|*|/|^|%%|%/%), relational (>|<|>=|<=|==|!=|%/%) and logic operators (&||), base R functions (e.g., abs(variable_name)), variables names (i.e., names(attTbl)) and class vectors (referred to as "classVector").

A combination of absolute and relative conditions can be used, but only one neighborhood condition per string is allowed. In order to avoid errors, it should always be possible to evaluate the conditions as TRUE or FALSE when substituting variable names with numeric values. All variables of interest should be addressed in the conditions argument by name.

Evaluation of conditions

The way conditions are evaluated depends on both the type of condition and the type of function used. When class contiguity is not being considered (i.e., functions without the argument nbs_of) absolute conditions can be used. Their evaluation is vectorized and exclude cells that have already been classified (i.e., which(!is.na(classVector))), unless the argument overwrite_class is TRUE. The function cond.reclass can evaluate absolute focal neighborhood conditions, but cannot evaluate directional neighborhood conditions (see functions cond.4.all and cond.reclass).

When class contiguity is being considered (i.e., functions with the argument nbs_of and anchor.seed), then both absolute and relative conditions can be used. Their evaluation is iterative. At each iteration, conditions are evaluated for the 8-neighbors of one of the cells classified as nbs_of. Neighbors that have already been classified are excluded from the evaluation (i.e., which(!is.na(classVector))), unless the argument overwrite_class is TRUE (see functions anchor.seed and cond.4.nofn). Some classification functions do not have a condition argument. These functions consider only class contiguity to classify raster cells (see functions anchor.cell, anchor.svo, reclass.nbs and classify.all).

Hereinafter the following definitions are adopted (see also the example section):

focal cell: cell included in the classes of the argument nbs_of whose neighbors are in evaluation;

nbs cell: one of the neighbors in evaluation;

focal neighborhood: when absolute focal neighborhood conditions are used, it includes nbs and its 8 neighbors; when relative focal neighborhood conditions are used, it only includes the 8 neighbors of nbs.

directional neighborhood: it consists of the intersection set of the focal cell neighbors with the focal neighborhood. When absolute focal neighborhood conditions are used, it includes nbs, but it does not include the focal cell; when relative focal neighborhood conditions are used, it includes the focal cell, but it does not include nbs.

Absolute conditions
This type of condition applies to all functions with a conditions argument. It compares between variables (including "classVector") and numeric values. When class contiguity is not considered, all cells meeting absolute conditions receive a classification number. When class contiguity is considered, all nbs meeting absolute conditions receive a classification number.

Examples of valid conditions: "variable_A > 1 & variable_B != 0"; "(variable_A^2 < 50 & variable_B == 0) | abs(variable_C) > 50".
Functions: anchor.seed, cond.4.all, cond.4.nofn and cond.reclass.

Absolute focal neighborhood conditions
This type of condition applies to the functions cond.4.nofn and cond.reclass. It compares each cell of focal neighborhood with a numeric value. The argument fn_perc controls the percentage of evaluations that have to be TRUE in order to assign a classification number to nbs. The function cond.4.nofn can consider a directional neighborhood instead of a focal neighborhood. This type of condition is flagged by a variable name followed by curly brackets (i.e., "variable_name{}"), where the curly brackets indicate the focal or directional neighborhood considered at each iteration.

Example of valid conditions: "variable_A{} > 1 & abs(variable_B) != 0".
Functions: cond.4.nofn and cond.reclass.

Focal cell conditions.
This type of condition applies only to functions considering class contiguity and with a conditions argument. It compares the value of nbs with the value of the focal cell. This type of condition is flagged by a variable name followed by square brackets (i.e., "variable_name[]"), where the square brackets indicate the focal cell.

Examples of valid conditions: "variable_A > variable_A[]"; "(variable_A > variable_A[] & variable_B{} < 10) | variable_C > 1". Note that the last example is a combination of absolute and focal cell conditions.
Functions: anchor.seed and cond.4.nofn.

Relative focal neighborhood conditions.
This type of condition applies only to the functions cond.4.nofn and cond.reclass. It compares each cell of focal neighborhood with nbs. The argument fn_perc controls the percentage of evaluations that have to be TRUE in order to assign a classification number to nbs. The function cond.4.nofn can consider a directional neighborhood instead of a focal neighborhood. This type of condition is flagged by a variable name followed by curly brackets (i.e., "variable_name{}"), where the curly brackets indicate the focal or directional neighborhood considered at each iteration.

Example of valid conditions: "variable_A > variable_A{}"; "(variable_A > variable_A{} & variable_B != variable_B[]) | variable_C > 1". Note that the last example is a combination of absolute, focal cell and focal neighborhood conditions.
Functions: cond.4.nofn and cond.reclass

See also

Examples

# EXAMPLES OF VALID AND INVALID CONDITIONS:

names_attTbl <- c("bathymetry", "slope")
cond <- "bathymetry>10"
conditions(names_attTbl, cond)
#> [1] "'Absolute' condition type(s) detected)."
cond <- "classVector != 1"
conditions(names_attTbl, cond)
#> [1] "'Class vector' condition type(s) detected)."
cond <- "bathymetry[]>10 & abs(slope{}) < 5"
conditions(names_attTbl, cond)
#> [1] "'Focal cell' AND 'Focal neighborhood' condition type(s) detected)."
if (FALSE) cond <- "thymetry[]>10 & abs(slpe{}) < 5"
conditions(names_attTbl, cond)
#> [1] "'Focal cell' AND 'Focal neighborhood' condition type(s) detected)."
cond <- "bathymetry[]>10 & | abs(slope{}) < 5"
conditions(names_attTbl, cond)
#> Error in conditions(names_attTbl, cond): Check: '|' -- Possible syntax error.
cond <- "bathymetry{}>10 & | abs(slope{}) < 5"
conditions(names_attTbl, cond)
#> Error in conditions(names_attTbl, cond): Only one 'neighborhood' condition allowed.

# EXAMPLES OF FOCAL CELL, NBS CELLS AND NEIGHBORHOODS:

# Matrix m mocking a raster of 3 rows and 4 columns
m <- matrix(1:12, nrow = 3, ncol = 4, byrow = TRUE)
m
#>      [,1] [,2] [,3] [,4]
#> [1,]    1    2    3    4
#> [2,]    5    6    7    8
#> [3,]    9   10   11   12

# FOCAL CELL
fc <- 6
fc
#> [1] 6

# NBS CELLS
nbs <- nbg8(3, 4)[[6]]
nbs
#> [1]  1  2  3  5  7  9 10 11

# CELL IN EVALUATION
nbs1 <- nbs[1]
nbs1
#> [1] 1

# FOCAL NEIGHBORHOOD
# Absolute:
ab_fn_nbs1 <- c(nbs1, nbg8(3, 4)[[1]])
ab_fn_nbs1
#> [1] 1 2 5 6

# Relative:
r_fn_nbs1  <- c(nbg8(3, 4)[[1]])
r_fn_nbs1
#> [1] 2 5 6

# DIRECTIONAL FOCAL NEIGHBORHOOD
# Absolute:
ab_dfn_nbs1 <- c(nbs1, intersect(c(nbg8(3, 4)[[1]]) , nbg8(3, 4)[[6]]))
ab_dfn_nbs1
#> [1] 1 2 5

# Relative:
r_dfn_nbs1  <- c(fc, intersect(c(nbg8(3, 4)[[1]]) , nbg8(3, 4)[[6]]))
r_dfn_nbs1
#> [1] 6 2 5