Eight Neighbors for Complete Cases

The function ngbList returns the list of 8-neighbors for all the cells of rstack with complete cases (i.e., cells without any missing value). Neighbors are identified by their cell numbers if the argument rNumb = FALSE or by their positions in the attribute table (i.e. row numbers) if the argument rNumb = TRUE and the argument attTbl != NULL (see attTbl).

The list of 8-neighbors is named. When rNumb = FALSE, names identify the cell number for which the neighborhood was computed. When rNumb = TRUE, names refers to row indices. For instance, the name "6" can be used to call the neighborhood of cell number 6 when rNumb = FALSE. However, when rNumb = TRUE, the name "6" can be used to call the neighborhood of the raster cell stored in the 6th row of the attribute table attTbl.

When the argument rNumb = TRUE, only neighbors with complete cases are considered, i.e. if a neighbor position correspond to a raster cell with one or more missing values it is not considered. Therefore, if a cell has all neighboring cells with missing values, the 8-neighbor vector of that cell will have zero-length.

ngbList(rstack, rNumb = FALSE, attTbl = NULL)

Arguments

rstack

Raster* object.

rNumb

logic, the 8-neighbors of each cell are identified by their row number in the attribute table (indicated by the argument attTbl). For instance, if cell 3 is located in row 2 it will be identified by the number 2.

attTbl

data.frame, the attribute table returned by the function attTbl. It is required only if the argument rNumb = TRUE

Value

Named list of integer vectors.

Details

The names of the list correspond to focal cells (as.character), the integer vectors within the list contain the 8-neighbors of each focal cell. When the argument rNumb = FALSE, both focal cells and neighbors are identified by cell numbers. When the argument rNumb = TRUE, both focal cells and neighbors are identified by row numbers.

Note

There is a correspondence between the indices of the attribute table (attTbl) and the indices of the 8-neighbors list. For instance, the first element in the 8-neighbors list corresponds to the neighbors of the cell whose values are stored in the first row of the attribute table.

See also

Examples

## r1, raster with cell numbers as values
r1 <- raster::raster(matrix(1:12, nrow = 3, ncol = 4, byrow = TRUE))

## r2, raster with missing values in cell numbers 1 and 2
r2 <- raster::raster(matrix(1, nrow = 3, ncol = 4, byrow = TRUE))
r2[c(1,2)] <- NA

ngbList(raster::stack(r1,r2))
#> $`3`
#> [1] 2 4 6 7 8
#> 
#> $`4`
#> [1] 3 7 8
#> 
#> $`5`
#> [1]  1  2  6  9 10
#> 
#> $`6`
#> [1]  1  2  3  5  7  9 10 11
#> 
#> $`7`
#> [1]  2  3  4  6  8 10 11 12
#> 
#> $`8`
#> [1]  3  4  7 11 12
#> 
#> $`9`
#> [1]  5  6 10
#> 
#> $`10`
#> [1]  5  6  7  9 11
#> 
#> $`11`
#> [1]  6  7  8 10 12
#> 
#> $`12`
#> [1]  7  8 11
#> 
matrix(1:12, nrow = 3, ncol = 4, byrow = TRUE)
#>      [,1] [,2] [,3] [,4]
#> [1,]    1    2    3    4
#> [2,]    5    6    7    8
#> [3,]    9   10   11   12