Geomorphic Units' Neighbors

Identify all the units adjacent to geomorphic units of interest. For each adjacent unit, individual borders are identified. Both function inputs and outputs use indices pointing to raster cells in the attribute table (see attTbl and ngbList with the argument index = TRUE).

obj.nbs(grp.bord, ngbList, only_grp = NULL, silent = FALSE)

Arguments

grp.bord

named list returned by the function obj.border. Each element of the list represents a geomorphic unit identified by the list's names, while the values of each element of the list consist of the indices pointing to the rows in the attribute table that correspond to the raster cells constituting the border.

ngbList

list, it has to contain the list of 8-neighbors of each cell in attTbl$Cell (see attTbl). This list has to be generated setting the argument of the function ngbList index = TRUE. (see ngbList).

only_grp

character vector, if geomorphic unit-IDs are provided as a character vector, then these are the only ones considered. If this argument is set as NULL, then all the geomorphic units included in the argument grp.bord are considered.

silent

logic, progress bar is not printed on the console.

Value

The function returns a named list. Each element of the list represent a geomorphic unit as identified by the list's names. Each element is constituted by a nested named list. The names of the nested lists are the IDs of all the adjacent geomorphic units. The values of each nested list consist of indices. These indices point to the rows in the attribute table (see attTbl) that correspond to the raster cells that mark the border between the geomorphic unit and the adjacent unit. The actual raster cells comprising borders can be extracted from the attribute table (e.g. attTbl$Cell[indices]) (see attTbl).