Set Anchor Cells from Local Minima or Local Maxima

Returns a vector of cell numbers at the locations of local minima or local maxima. These cells can be used as anchor cells in other scapesClassification functions.

anchor.seed(
  attTbl,
  ngbList,
  rNumb = FALSE,
  class = NULL,
  cond.filter = NULL,
  cond.seed,
  sort.seed = NULL,
  sort.col = NULL,
  cond.growth = NULL,
  lag.growth = Inf,
  cond.isol = NULL,
  lag.isol = 1,
  classVector = NULL,
  saveRDS = NULL,
  overWrite = FALSE,
  isolationClass = FALSE,
  silent = FALSE
)

Arguments

attTbl

data.frame, the attribute table returned by the function attTbl.

ngbList

list, it has to contain the list of 8-neighbors of each cell in attTbl$Cell (see ngbList).

rNumb

logic, ngbList contain the neighbors index position in the attribute table (see ngbList).

class

numeric, the numeric class to attribute to local minima or maxima. If NULL then each local minima or maxima is classified with a different number starting from one. Buffers have the same classification number as the local minima or maxima to which they refer.

cond.filter

character string, the conditions to define for what cells the arguments cond.seed, cond.growth and cond.isol have to be evaluated. It can be NULL. Absolute conditions can be used (see conditions).

cond.seed

character string, the conditions to identify local maxima or minima. The condition should be similar to "variable_x == max(variable_x)" or "variable_x == min(variable_x)". It cannot be NULL.

sort.seed

character, sort seeds based on column values. If "max" seeds are evaluated from the maximum to the minimum. If "min" seeds are evaluated from the minimum to the maximum.

sort.col

character, the column name in the attTbl on which the sorting in based on.

cond.growth

character string, the conditions to define a buffer around local maxima or minima. It can be NULL. Absolute and focal cell conditions can be used (see conditions).

lag.growth

numeric, it defines the lag on which focal cell conditions in cond.growth are evaluated (see conditions).

cond.isol

character string, the conditions to define how one local maxima or minima is isolated from another. It can be NULL. Absolute and focal cell conditions can be used (see conditions).

lag.isol

numeric, it defines the lag on which focal cell conditions in cond.isol are evaluated (see conditions).

classVector

numeric vector, if any of the condition arguments refers to a previous classification, then this classification have to be referred as "classVector".

saveRDS

filename, if a file name is provided save the anchor cell vector as an RDS file.

overWrite

logic, if the RDS names already exist, existing files are overwritten.

isolationClass

logic, return cells meeting isolation conditions (isolator cells) classified as '-1'.

silent

logic, progress is not printed on the console.

Value

Classification vector.

Details

This function implements an algorithm to identify local maxima or minima based on a set of conditions. It is possible to identify local maxima or minima considering only a subset of cells using the argument cond.filter.

The algorithm starts identifying the cell(s) meeting the conditions of the argument cond.seed. This cell(s) can be referred to as seed cell(s). If a cond.growth argument is provided, all cells around the seed meeting the conditions of cond.growth are also classified as seed. The algorithm will continue looking for new seed cells as long as among the neighbors of the newly classified seed cells there are cells that meet the conditions of cond.growth. In the next step, cells adjacent to seed cells meeting the conditions of the argument cond.isol are classified as isolator cells. As for seed cells, the algorithm will continue looking for isolator cells as long as among the neighbors of the newly classified isolator cells there are cells that meet the conditions of cond.isol. When a cell is classified either as seed or as isolator, then it is ignored in all successive iterations of the algorithm. The cond.seed, cond.growth, cond.isol process is repeated until all cells are classified either as seed or as isolator.

The evaluation of cond.growth and cond.isol follows a node structure. Let us consider cond.growth as an example and refer to 'the conditions of the cond.growth argument' simply as cond. At level_0 there is a single node, a cell previously classified as seed. Each cell adjacent to the node in level_0 that evaluate positively to cond represents a node at level_1. Suppose that at level_1 there are one or more nodes. The algorithm then considers the first of these nodes and looks for neighbor cells evaluating positively to cond. If one or more cells do, then the first node of level_1 is set as referenceNode for level_2. The algorithm then evaluates if one or more cells adjacent to the first node in level_2 evaluate positively to cond. If any does, then the first node of level_2 is set as referenceNode for level_3. The algorithm continues in this way until level_n where none of the adjacent cells evaluate positively to cond. At this point the algorithm considers cells adjacent to the second node at level_n-1. If any cell has a positive evaluation, then second node in level_n-1 is set as referenceNode for level_n and the algorithm continues as described above. On the contrary, if none of the cells evaluate positively to cond, the algorithm will remain at at level_n-1 and will consider the next node. If none of the nodes at level_n-1 has any neighboring cell evaluating positively to cond, then the algorithm will move down to level_n-2. The evaluation of cond.growth will stop once the algorithm returns to level_0.

This implementation allows to track back to level_0 all the connected cells having a positive evaluation. When conditions relative to the focal cell are considered (see conditions), the arguments lag.growth and lag.isol determine to which referenceNode newly evaluated cells have to be compared to. A lag of 1 indicates that the conditions at level_n have to be evaluated considering the referenceNode at level_n-1. A lag set as Inf indicates that at any level the conditions are evaluated considering the initial node cell at level_0. If the lag is greater than the level at which the evaluation occurs, then the conditions are evaluated considering the initial node cell at level_0 (e.g., level_1, lag = 2).

See also