Test conditions for all cells

Evaluate conditions for unclassified cells and classify them if conditions are true.

cond.4.all(
  attTbl,
  conditions,
  classVector = NULL,
  class,
  overwrite_class = FALSE
)

Arguments

attTbl

data.frame, the attribute table returned by the function attTbl.

conditions

character string, the conditions a cell have to meet to be classified as indicated by the argument class. If there is a classVector input, the classification number is only assigned to classVector NA-cells.

classVector

numeric vector, if provided, it defines the cells in the attribute table that have already been classified and that have to be ignored by the function (unless the argument overwrite_class = TRUE).

class

numeric, the classification number to assign to all cells that meet the function conditions.

overwrite_class

logic, if there is a classVector input, reclassify cells that were already classified and that meet the function conditions.

Value

Update classVector with the new cells that were classified by the function. If there is no classVector input, the function returns a new class vector.

Details

The function evaluates the conditions of the argument conditions for all unclassified cells (i.e., classVector NA-cells).

Cells that meet the function conditions are classified as indicted by the argument class.

Absolute test cell conditions can be used (see conditions).

See also

Examples

# NOT RUN
if (FALSE) {
# LOAD LIBRARIES
library(raster)
library(scapesClassification)

# LOAD THE DUMMY RASTER
r <- list.files(system.file("extdata", package = "scapesClassification"),
                pattern = "dummy_raster\\.tif", full.names = T)
r <- raster(r)

# COMPUTE THE ATTRIBUTE TABLE
at <- attTbl(r, "dummy_var")

# COMPUTE THE LIST OF NEIGBORHOODS
nbs <- ngbList(r)

# COMPUTE A NEW CLASS VECTOR (PLOT 1)
################################################################################
# conditions: "dummy_var == 1"
cv1   <- cond.4.all(attTbl = at, conditions = "dummy_var == 1", class = 10)

unique(cv1) # one class (class 10)
################################################################################

# UPDATE THE PREVIOUS CLASS VECTOR (PLOT 2)
################################################################################
# conditions: "dummy_var <= 3"
cv2   <- cond.4.all(attTbl = at, conditions = "dummy_var <= 3", class = 1,
                    classVector = cv1) # previous class vector

unique(cv2) # two classes (class 10 and class 1)
################################################################################

# CONVERT THE CLASS VECTOR INTO A RASTER AND PLOT
r_cv1 <- cv.2.rast(r, at$Cell, classVector = cv1)
r_cv2 <- cv.2.rast(r, at$Cell, classVector = cv2)

# PLOTs
par(mar = c(1, 0, 2.3, 0), par(mfrow=c(1,2)))

# plot 1
plot(r_cv1, axes=FALSE, box=FALSE, legend = FALSE, colNA="#818792",
     col="#78b2c4", main = "New class vector\n 'dummy_var == 1'")

# show r values
text(r)

# plot 2
plot(r_cv2, axes=FALSE, box=FALSE, legend = FALSE, colNA="#818792",
     col=c("#cfad89", "#78b2c4"), main = "Update class vector\n 'dummy_var <= 3'")

# show r values
text(r)
}