Geomorphic Units' Borders

Takes as input a list of elements representing geomorphic units and identify the border of each unit.

obj.border(group, ngbList, silent = FALSE)

Arguments

group

named list, each element represents a geomorphic unit. Each unit is identified by the indices pointing to the raster cells in the attribute table that compose that unit (see attTbl).

ngbList

list, it has to contain the list of 8-neighbors of each cell in attTbl$Cell (see attTbl). This list has to be generated setting the argument of the function ngbList index = TRUE (see ngbList).

silent

logic, progress bar is not printed on the console.

Value

The function returns a named list. Each element of the list represents a geomorphic unit as identified by the list's names, while the values of each element of the list consist of indices. These indices point to the rows in the attribute table that correspond to the raster cells comprising the borders. The actual raster cells comprising borders can be extracted from the attribute table (e.g. attTbl$Cell[indices]) (see attTbl).

Note

Note that group has to be a named list whose names correspond to geomorphic unit-IDs.

See also